requirements.txt
absl-py==2.1.0
aiobotocore>=2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.3.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
boto3==1.34.149
botocore==1.34.149
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.1
colorama==0.4.6
colorclass==2.2.2
comm==0.2.2
configobj==5.0.8
constantly==23.10.4
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
docopt==0.6.2
dtw-python==1.5.1
EMD-signal==1.6.0
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.65.1
h5py==3.11.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.2.0
incremental==24.7.0
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
keras==3.4.1
keyring==25.2.1
kiwisolver==1.4.5
launchpadlib==2.0.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
libclang==18.1.1
lit==18.1.8
llvmlite==0.43.0
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
multitasking==0.0.11
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.2.1.18
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
nvidia-tensorrt==99.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathos==0.3.2
peewee==3.17.6
pexpect==4.9.0
pickleshare==0.7.5
pika==1.3.2
pillow==10.4.0
pip-upgrader==1.4.15
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.23.0
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyHamcrest==2.1.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
pyrsistent==0.20.0
pyserial==3.5
python-dateutil==2.9.0
pytorch-ignite==0.6.0.dev20240726
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
randomname==0.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
s3fs==2024.6.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
SecretStorage==3.3.3
service-identity==24.1.0
setuptools==71.1.0
shap==0.46.0
simplejson==3.19.2
six==1.16.0
slicer==0.0.8
smmap==5.0.1
soupsieve==2.5
ssh-import-id==5.11
st-ant-tree==0.0.12
st-files-connection==0.1.0
stack-data==0.6.3
streamlit==1.39.0
streamlit-tree-select==0.0.5
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorrt==10.2.0.post1
tensorrt-cu12==10.2.0.post1
tensorrt-cu12-bindings==10.2.0.post1
tensorrt-cu12-libs==10.2.0.post1
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
triton==3.0.0
Twisted==24.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wadllib==1.3.6
watchdog==4.0.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
yarg==0.1.9
yarl==1.9.4
yfinance==0.2.41
zipp==3.19.2
zope.interface==6.4
